+++ src/csstock/main.cpp
@@ -24,8 +24,8 @@ public:
   void set_apis(SignalApis);
   void load_state(const std::string &indir);
   void save_state(const std::string &outdir);
-  void on_sod(uint32_t date, const SodEvent *ev);
-  void on_eod(uint32_t date);
+  void on_sod(const SodEvent *ev);
+  void on_eod(const EodEvent *ev);
 
   void on_cs_snapshot(const CsSnapshotEvent *ev);
 
@@ -36,7 +36,7 @@ private:
 
 // function definitions
 
-Signal::Signal() { on_sod(0, nullptr); }
+Signal::Signal() {}
 
 void Signal::initialize(const Config *root) {}
 
@@ -50,11 +50,8 @@ void Signal::save_state(const std::string &outdir) {
   wllog_info("saving state to dir {}\n", outdir);
 }
 
-void Signal::on_sod(uint32_t date, const SodEvent *ev) {
+void Signal::on_sod(const SodEvent *ev) {
   m_cnt = 0;
-  if (!ev) {
-    return;
-  }
   // NOTE:
   // for now in cross-sectional mode, we get the full list of stock names
   // on start of each day
@@ -71,7 +68,7 @@ void Signal::on_sod(uint32_t date, const SodEvent *ev) {
   }
 }
 
-void Signal::on_eod(uint32_t date) {
+void Signal::on_eod(const EodEvent *ev) {
   wllog_info("{} updates received\n", m_cnt);
 }
 
@@ -126,14 +123,14 @@ static const SignalOps my_ops = {
       ptr->save_state(outdir);
     },
 
-    .on_sod = [](void *hdl, uint32_t date, const SodEvent *ev) -> void {
+    .on_sod = [](void *hdl, const SodEvent *ev) -> void {
       auto *ptr = reinterpret_cast<Signal *>(hdl);
-      ptr->on_sod(date, ev);
+      ptr->on_sod(ev);
     },
 
-    .on_eod = [](void *hdl, uint32_t date) -> void {
+    .on_eod = [](void *hdl, const EodEvent *ev) -> void {
       auto *ptr = reinterpret_cast<Signal *>(hdl);
-      ptr->on_eod(date);
+      ptr->on_eod(ev);
     },
 
     .on_cs_snapshot = [](void *hdl, const CsSnapshotEvent *ev) -> void {