+++ src/multitickers/main.cpp
@@ -21,8 +21,8 @@ public:
 
   void initialize(const Config *root);
   void set_apis(SignalApis);
-  void on_sod(uint32_t date, const SodEvent *ev);
-  void on_eod(uint32_t date);
+  void on_sod(const SodEvent *ev);
+  void on_eod(const EodEvent *ev);
 
   void on_snapshot(const SnapshotEvent *ev);
 
@@ -33,24 +33,21 @@ private:
 
 // function definitions
 
-Signal::Signal() { on_sod(0, nullptr); }
+Signal::Signal() {}
 
 void Signal::initialize(const Config *root) {}
 
 void Signal::set_apis(SignalApis apis) { m_apis = apis; }
 
-void Signal::on_sod(uint32_t date, const SodEvent *ev) {
+void Signal::on_sod(const SodEvent *ev) {
   m_cnt = 0;
-  if (!ev) {
-    return;
-  }
   // NOTE:
   // for now in cross-sectional mode, we get the full list of stock names
   // on start of each day
   wllog_info("ins_nr={}\n", ev->ins_nr);
 }
 
-void Signal::on_eod(uint32_t date) {
+void Signal::on_eod(const EodEvent *ev) {
   wllog_info("{} updates received\n", m_cnt);
 }
 
@@ -74,14 +71,14 @@ static SignalOps my_ops = {
       ptr->set_apis(apis);
     },
 
-    .on_sod = [](void *hdl, uint32_t date, const SodEvent *ev) -> void {
+    .on_sod = [](void *hdl, const SodEvent *ev) -> void {
       auto *ptr = reinterpret_cast<Signal *>(hdl);
-      ptr->on_sod(date, ev);
+      ptr->on_sod(ev);
     },
 
-    .on_eod = [](void *hdl, uint32_t date) -> void {
+    .on_eod = [](void *hdl, const EodEvent *ev) -> void {
       auto *ptr = reinterpret_cast<Signal *>(hdl);
-      ptr->on_eod(date);
+      ptr->on_eod(ev);
     },
 
     .on_snapshot = [](void *hdl, const SnapshotEvent *ev) -> void {